Add support for parameterized Cache Manager queries fix warnings in CentOS 5.3 amd64

+++ src/mgr/Inquirer.cc
@@ -161,7 +161,7 @@ Mgr::Inquirer::applyQueryParams(const Ipc::StrandCoords& aStrands, const QueryPa
             IntParam* param = dynamic_cast<IntParam*>(workersParam.getRaw());
             if (param != NULL && param->type == QueryParam::ptInt) {
                 const std::vector<int>& workers = param->value();
-                for (size_t i = 0; i < aStrands.size(); ++i)
+                for (int i = 0; i < (int)aStrands.size(); ++i)
                 {
                     if (std::find(workers.begin(), workers.end(), i + 1) != workers.end())
                         strands.push_back(aStrands[i]);
